requirements.txt
pandas
streamlit
requests
fpdf2
matplotlib